# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lucy API",
    "version": "1.0.0"
  },
  "paths": {
    "/conversationIds": {
      "get": {
        "summary": "Get a list of conversation IDs",
        "parameters": [
          {
            "name": "agentName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of conversation IDs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing agentName or accountName)"
          }
        }
      },
      "put": {
        "description": "Rename conversationId",
        "parameters": [
          {
            "name": "agentName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "existingId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request (missing existingId or newId)"
          }
        }
      }
    },
    "/computeConversations": {
      "post": {
        "summary": "Get the prompt that would be used given an agentName, accountName, selectType and query text",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "agentName": {
                    "type": "string"
                  },
                  "accountName": {
                    "type": "string"
                  },
                  "selectType": {
                    "type": "string"
                  },
                  "query": {
                    "type": "string"
                  },
                  "conversationId": {
                    "type": "string"
                  }
                },
                "required": [
                  "agentName",
                  "accountName",
                  "selectType",
                  "query",
                  "conversationid"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A list of prompt elements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "role": {
                        "type": "string"
                      },
                      "content": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "role",
                      "content"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing agentName, accountName, selectType or query)"
          }
        }
      }
    },
    "/ask": {
      "post": {
        "summary": "Ask a question",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string"
                  },
                  "agentName": {
                    "type": "string"
                  },
                  "accountName": {
                    "type": "string"
                  },
                  "conversationId": {
                    "type": "string"
                  }
                },
                "required": [
                  "question",
                  "agentName",
                  "accountName",
                  "conversationId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI assistant's response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing question, agentName, or accountName)"
          }
        }
      }
    },
    "/agents": {
      "get": {
        "summary": "Get a list of agents",
        "responses": {
          "200": {
            "description": "A list of agent objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "agentName": {
                        "type": "string"
                      },
                      "seed_conversation": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "role": {
                              "type": "string"
                            },
                            "content": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "role",
                            "content"
                          ]
                        }
                      },
                      "language_code": {
                        "type": "string"
                      },
                      "select_type": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "agentName",
                      "seed_conversation",
                      "language_code",
                      "select_type"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/prompts": {
      "post": {
        "summary": "Get the prompt that would be used given an agentName, accountName, selectType and query text",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "agentName": {
                    "type": "string"
                  },
                  "accountName": {
                    "type": "string"
                  },
                  "selectType": {
                    "type": "string"
                  },
                  "query": {
                    "type": "string"
                  },
                  "conversationId": {
                    "type": "string"
                  }
                },
                "required": [
                  "agentName",
                  "accountName",
                  "selectType",
                  "query",
                  "conversationid"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A list of prompt elements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "role": {
                        "type": "string"
                      },
                      "content": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "role",
                      "content"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing agentName, accountName, selectType or query)"
          }
        }
      },
      "get": {
        "summary": "Get one or more prompts",
        "parameters": [
          {
            "name": "agentName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "conversationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of prompt objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Prompt"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a prompt",
        "parameters": [
          {
            "name": "agentName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "prompt": {
                    "$ref": "#/components/schemas/Prompt"
                  }
                },
                "required": [
                  "id",
                  "prompt"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated prompt object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prompt"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing accountName or prompt object)"
          }
        }
      },
      "delete": {
        "summary": "Delete a prompt",
        "parameters": [
          {
            "name": "agentName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prompt successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing accountName or id)"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Prompt": {
        "type": "object",
        "properties": {
          "conversation": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "utc_timestamp": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "content",
                "role",
                "utc_timestamp"
              ]
            }
          },
          "conversationId": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "total_chars": {
            "type": "integer"
          },
          "utc_timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "conversation",
          "conversationId",
          "id",
          "keywords",
          "total_chars",
          "utc_timestamp"
        ]
      }
    }
  }
}